refactor: refactoring file naming and divide css file
