+++ src/App.vue
@@ -1,19 +1,19 @@
-<script setup lang="ts">
-import { RouterView, RouterLink } from 'vue-router'
-import NavBar from '@/common/NavBar.vue';
-</script>
-
 <template>
   <div class="container">
     <header>
-      <NavBar />
+      <Navbar/>
     </header>
-    <div>
-      <RouterView />
+    <div class="content">
+      <RouterView/>
     </div>
   </div>
 </template>
 
-<style scoped>
+<script setup lang="ts">
+import {RouterView} from 'vue-router'
+import Navbar from '@/common/Navbar.vue';
+</script>
+
+<style>
 @import "./assets/css/main.css";
 </style>
