--- src/assets/css/conditionbar.css
@@ -1,90 +1,0 @@
-.condition {
-    display: flex;
-    align-items: flex-end;
-    width: 600px;
-    height: 40px;
-    gap: 20px;
-}
-
-.search {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    flex: 1 0 0;
-}
-
-.search-input-container {
-    display: flex;
-    align-items: center;
-    gap: 0.625rem;
-    flex: 1 0 0;
-    align-self: stretch;
-}
-
-.search-input {
-    display: flex;
-    align-items: center;
-    gap: 0.625rem;
-    flex: 1 0 0;
-    align-self: stretch;
-    padding-left: 10px;
-
-    color: #000;
-    text-align: left;
-    font-family: Open Sans;
-    font-size: 15px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 140%; /* 1.75rem */
-
-    border-radius: 5px;
-    border: 1px solid #7C7C7C;
-}
-
-.search-input:focus {
-    outline: none !important;
-    border: 1px solid #696CFF;
-}
-
-.sort {
-    display: flex;
-    width: 200px;
-    justify-content: center;
-    align-items: center;
-    gap: 20px;
-    align-self: stretch;
-}
-
-a {
-    color: #7C7C7C;
-    text-decoration: none;
-}
-
-.sort-option {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-    flex: 1 0 0;
-    align-self: stretch;
-
-    border-radius: 5px;
-    border: 1px solid #7C7C7C;
-
-    color: #7C7C7C;
-    text-align: center;
-    font-family: Open Sans;
-    font-size: 15px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 140%;
-}
-
-.selected * {
-    color: #696CFF;
-    font-weight: 700;
-}
-
-.sort-option.selected {
-    border: 1px solid #696CFF;
-}
