--- src/assets/css/info.css
@@ -1,6 +1,0 @@
-.info {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 74px;
-}
