+++ src/assets/css/main.css
@@ -1,5 +1,22 @@
 .container {
     display: flex;
+    flex-direction: row;
     align-items: center;
+    width: 100%;
     gap: 50px;
+}
+
+.content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 74px;
+}
+
+.content-header {
+    display: flex;
+    align-items: flex-end;
+    width: 600px;
+    height: 40px;
+    gap: 20px;
 }
