+++ src/assets/css/search.css
@@ -1,0 +1,39 @@
+.search {
+    display: flex;
+    height: 100%;
+    align-items: center;
+    flex: 1 0 0;
+}
+
+.search-input-container {
+    display: flex;
+    align-items: center;
+    gap: 0.625rem;
+    flex: 1 0 0;
+    align-self: stretch;
+}
+
+.search-input {
+    display: flex;
+    align-items: center;
+    gap: 0.625rem;
+    flex: 1 0 0;
+    align-self: stretch;
+    padding-left: 10px;
+
+    color: #000;
+    text-align: left;
+    font-family: Open Sans;
+    font-size: 15px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 140%; /* 1.75rem */
+
+    border-radius: 5px;
+    border: 1px solid #7C7C7C;
+}
+
+.search-input:focus {
+    outline: none !important;
+    border: 1px solid #696CFF;
+}
