+++ src/assets/css/sort.css
@@ -1,0 +1,42 @@
+.sort {
+    display: flex;
+    width: 200px;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+    align-self: stretch;
+}
+
+a {
+    color: #7C7C7C;
+    text-decoration: none;
+}
+
+.sort-option {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    flex: 1 0 0;
+    align-self: stretch;
+
+    border-radius: 5px;
+    border: 1px solid #7C7C7C;
+
+    color: #7C7C7C;
+    text-align: center;
+    font-family: Open Sans;
+    font-size: 15px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 140%;
+}
+
+.selected * {
+    color: #696CFF;
+    font-weight: 700;
+}
+
+.sort-option.selected {
+    border: 1px solid #696CFF;
+}
