+++ src/components/TableView.vue
@@ -1,0 +1,34 @@
+<template>
+  <table>
+    <thead>
+    <th v-for="(header, index) in tableHeaders" :key="index">{{ header }}</th>
+    </thead>
+    <tbody v-for="(info, index) in tableInfos" :key="index">
+    <td v-for="property in tableProperties" :key="property">
+      <div v-if="(property === 'projectIsAuthorized') && info[property] === true">
+        <button class="project-auth-btn" type="button" disabled>인증</button>
+      </div>
+      <div v-else-if="(property === 'projectIsAuthorized') && info[property] === false">
+        <button class="project-unauth-btn" type="button" @click="authorizeProject(info.projectId)">미인증</button>
+      </div>
+      <div v-else>{{ info[property] }}</div>
+    </td>
+    </tbody>
+  </table>
+</template>
+
+<script lang="ts">
+export default {
+  name: 'TableInfo',
+  props: {
+    tableHeaders: Array<string>,
+    tableInfos: Array,
+    tableProperties: Array<string>,
+    authorizeProject: Function
+  }
+};
+</script>
+
+<style scoped>
+@import "@/assets/css/tableview.css";
+</style>
