+++ src/main.ts
@@ -1,6 +1,6 @@
-import "./assets/css/main.css";
-import { createApp } from "vue";
-import { createPinia } from "pinia";
+// import "./assets/css/main.css";
+import {createApp} from "vue";
+import {createPinia} from "pinia";
 
 import App from "./App.vue";
 import router from "./router";
